feat(portfolio): replace bespoke slider with react-compare-slider (#72)
- Extract portfolio data discovery into lib/portfolio.ts (PortfolioItem type + getPortfolioItems)
- Add CompareSlider client wrapper around ReactCompareSlider
- Add CompareSliderDynamic to prevent SSR hydration mismatch
- Update Portfolio.tsx to use new lib and components
- Remove setupSlider (~290 lines) from Scripts.tsx
- Remove Before/After Slider CSS block from globals.css
- Add portfolio-compare-slider smoke test (renders, keyboard a11y, axe)
