+++ web/app/globals.css
@@ -493,81 +493,6 @@ body {
     background-size: cover;
 }
 
-/* Before/After Slider */
-.comparison-slider {
-    position: relative;
-    overflow: hidden;
-    cursor: col-resize;
-}
-
-.comparison-slider img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    display: block;
-}
-
-.comparison-slider .after-image {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 50%;
-    height: 100%;
-    overflow: hidden;
-}
-
-.comparison-slider .slider-handle {
-    position: absolute;
-    top: 0;
-    left: 50%;
-    width: 3px;
-    height: 100%;
-    background: linear-gradient(180deg, #D4AF37, #FFD700);
-    transform: translateX(-50%);
-    pointer-events: none;
-}
-
-.comparison-slider .slider-handle::before {
-    content: ' ';
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    width: 44px;
-    height: 44px;
-    border: 2px solid #D4AF37;
-    border-radius: 50%;
-    background: rgba(26, 31, 46, 0.8);
-    transform: translate(-50%, -50%);
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 0 10px;
-    box-sizing: border-box;
-    box-shadow: 0 8px 30px rgba(0,0,0,0.35), 0 0 0 2px rgba(255,255,255,0.05) inset;
-}
-.comparison-slider .slider-handle::after {
-    content: '↔';
-    color: #ffffff;
-    font-size: 1.2rem;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
-
-/* Visible focus treatment so keyboard users can track handle position */
-.comparison-slider .slider-handle:focus-visible {
-    outline: none;
-}
-
-.comparison-slider .slider-handle:focus-visible::before {
-    box-shadow:
-        0 0 0 3px rgba(212, 175, 55, 0.7),
-        0 0 0 6px rgba(26, 31, 46, 0.9),
-        0 8px 30px rgba(0, 0, 0, 0.35),
-        0 0 0 2px rgba(255, 255, 255, 0.05) inset;
-}
-
 /* Scroll progress bar moved to framer-motion component styles */
 
 /* WhatsApp Widget */
