+++ web/components/CompareSliderDynamic.tsx
@@ -1,0 +1,6 @@
+"use client"
+import dynamic from "next/dynamic"
+
+const CompareSlider = dynamic(() => import("./CompareSlider"), { ssr: false })
+
+export default CompareSlider
