+++ web/package.json
@@ -31,6 +31,7 @@
     "lucide-react": "0.539.0",
     "next": "^16.1.6",
     "react": "19.2.4",
+    "react-compare-slider": "^4.0.0",
     "react-dom": "19.2.4",
     "react-email": "4.2.8",
     "resend": "^6.0.1",
